Endpoint to get endorsed items with multiple suppliers assigned added

+++ src/main/java/com/logistics/supply/repository/PaymentDraftRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentDraftRepository extends JpaRepository<PaymentDraft, Long> {
+public interface PaymentDraftRepository extends JpaRepository<PaymentDraft, Integer> {
 
 }
